+++ src/imio/news/core/browser/bring_news_into_news_folders/news_folders.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+
+from imio.news.core.viewlets.news import (
+    user_is_contributor_in_entity_which_authorize_to_bring_news,
+)
 from imio.smartweb.common.utils import get_vocabulary
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api
@@ -11,8 +15,6 @@ from z3c.form import form
 from z3c.form.button import buttonAndHandler
 from plone.supermodel import model
 
-import transaction
-
 
 class IBringNewsIntoNewsFoldersForm(model.Schema):
     """ """
@@ -20,9 +22,10 @@ class IBringNewsIntoNewsFoldersForm(model.Schema):
     directives.widget(
         "news_folders",
         AjaxSelectFieldWidget,
-        source="imio.news.vocabulary.UserNewsFolders",
+        vocabulary="imio.news.vocabulary.UserNewsFolders",
         pattern_options={"multiple": True},
     )
+    directives.write_permission(news_folders="cmf.SetOwnProperties")
     news_folders = schema.List(
         title=_("Available news folders"),
         value_type=schema.Choice(source="imio.news.vocabulary.UserNewsFolders"),
@@ -38,17 +41,26 @@ class BringNewsIntoNewsFoldersForm(AutoExtensibleForm, form.Form):
     enable_autofocus = False
     label = _("Add/Remove news folder(s)")
 
+    def update(self):
+        super(BringNewsIntoNewsFoldersForm, self).update()
+        if user_is_contributor_in_entity_which_authorize_to_bring_news is False:
+            api.portal.show_message(
+                _("You don't have rights to access this page."), self.request
+            )
+            self.request.response.redirect(self.context.absolute_url())
+            return False
+
     def updateWidgets(self):
-        news_folders_to_display = []
-        vocabulary = get_vocabulary("imio.news.vocabulary.UserNewsFolders")
-        # Loop to display only news folders where user has the permission (ex : to remove these news folders out of news)
-        for news_folder_uid in self.context.selected_news_folders:
-            if vocabulary.by_token.get(news_folder_uid) is None:
-                # user can't remove this news folders because he has no permission on it so we don't display it
-                pass
-            news_folders_to_display.append(news_folder_uid)
-        self.fields["news_folders"].field.default = news_folders_to_display
         super(BringNewsIntoNewsFoldersForm, self).updateWidgets()
+        selectedItems = {}
+        self.selectedUID = []
+        vocabulary = get_vocabulary("imio.news.vocabulary.UserNewsFolders")
+
+        for term in vocabulary:
+            if term.value in self.context.selected_news_folders:
+                self.selectedUID.append(term.value)
+                selectedItems[term.value] = term.title
+        self.widgets["news_folders"].value = ";".join(self.selectedUID)
 
     @buttonAndHandler(_("Submit"))
     def handle_submit(self, action):
@@ -56,27 +68,25 @@ class BringNewsIntoNewsFoldersForm(AutoExtensibleForm, form.Form):
         if errors:
             self.status = self.formErrorsMessage
             return
-        if len(data.get("news_folders")) < len(
-            self.fields["news_folders"].field.default
-        ):
-            # we want to remove news folder(s) out of this news
+        if len(data.get("news_folders")) < len(self.selectedUID):
+            # we want to remove news folder(s) out of this event
             news_folders_to_remove = list(
-                set(self.fields["news_folders"].field.default)
-                - set(data.get("news_folders"))
+                set(self.selectedUID) - set(data.get("news_folders"))
             )
             for news_folder in news_folders_to_remove:
                 self.context.selected_news_folders.remove(news_folder)
             success_message = _("News folder(s) correctly removed.")
         else:
-            # we want to add news folder(s) in this news
-            for new_news_folder in data.get("news_folders"):
-                if new_news_folder not in self.context.selected_news_folders:
-                    self.context.selected_news_folders.append(new_news_folder)
+            # we want to add an news folder in this event
+            for news_folder in data.get("news_folders"):
+                if news_folder not in self.context.selected_news_folders:
+                    self.context.selected_news_folders.append(news_folder)
             success_message = _("News folder(s) correctly added.")
-        self.context.reindexObject(idxs=["selected_news_folders"])
-        transaction.commit()
+
+        self.context.reindexObject()
         self.status = success_message
         api.portal.show_message(_(self.status), self.request)
+
         self.request.response.redirect(self.context.absolute_url())
 
     @button.buttonAndHandler(_("Cancel"))